Program reports exit code. Colour on pointer types fixed.

+++ support/msRgb.h
@@ -64,16 +64,16 @@ public:
 
 	m_inline Rgb interpolated(int _percent, Rgb _c = black) const
 	{
-		return fromRgb(	(r() * (100 - _percent) + _c.r() * _percent) / 100,
-						(g() * (100 - _percent) + _c.g() * _percent) / 100,
-						(b() * (100 - _percent) + _c.b() * _percent) / 100,
-						(a() * (100 - _percent) + _c.a() * _percent) / 100);
+		return fromRgb(	((int)r() * (100 - _percent) + (int)_c.r() * _percent) / 100,
+						((int)g() * (100 - _percent) + (int)_c.g() * _percent) / 100,
+						((int)b() * (100 - _percent) + (int)_c.b() * _percent) / 100,
+						((int)a() * (100 - _percent) + (int)_c.a() * _percent) / 100);
 	}
 	m_inline Rgb interpolated(int _percent, uchar _alpha, Rgb _c = black) const
 	{
-		return fromRgb(	(r() * (100 - _percent) + _c.r() * _percent) / 100,
-						(g() * (100 - _percent) + _c.g() * _percent) / 100,
-						(b() * (100 - _percent) + _c.b() * _percent) / 100, _alpha);
+		return fromRgb(	((int)r() * (100 - _percent) + (int)_c.r() * _percent) / 100,
+						((int)g() * (100 - _percent) + (int)_c.g() * _percent) / 100,
+						((int)b() * (100 - _percent) + (int)_c.b() * _percent) / 100, _alpha);
 	}
 
 	static const Rgb black;
